Add in conventions info from HACKING
We can't really talk about how to contribute to Xapian without
being able to point to our conventions. Pointing at HACKING isn't
great because it's long and (unless you render it) difficult to
navigate.

While doing this, I've improved some of the rst markup, and also
sphinx-ified some of it. That's not completely ideal, because it
will make it harder to apply changes from HACKING until we replace
it in a future release with this developers' guide, but it makes
the output much nicer and more readable.

Update Olly's copyright assertion to cover everything from 2006
onwards. Although HACKING started long before then, most of
the useful information is much younger.