Remove/update Python2 references
For the dev guide, we're largely focussed on building the development
version from git master which no longer supports Python2.  Also the
package naming has changed in recent Debian and Ubuntu and remaining
Python2 packages now seem to be python2-XXX instead of python-XXX so
unhelpfully we can't give a list of packages which works across all
releases still in support.